initial code edit
Used CPW_WeedsTrtmnt_WebMap as a template for file structure and code.  Using code from ESRI tutorial as a jumping off point to create a thin client web  mapping application.  This method allows me to conficure popups, symbology, base map, and links to data sources on AGOL so that this funcionality does not have to be handled by the client allowing for thinner clients to use with less speed issues.